fix(metadata): correct data type casting in completion score macro and update required creator metadata format

+++ orchestration/dags/data_gcp_dbt/macros/function_utils/completion_score.sql
@@ -11,13 +11,13 @@
     round(
         1 - (
             (
-                cast({{ film_titre_format_support }} as float64) * 1.0
-                + cast({{ absence_image }} as float64) * 2.0
-                + cast({{ absence_video }} as float64) * 1.0
-                + cast({{ absence_description }} as float64) * 1.5
-                + cast({{ livre_cd_vinyle_artiste_manquant }} as float64) * 1.0
-                + cast({{ livre_cd_vinyle_gtl_manquant }} as float64) * 1.0
-                + cast({{ livre_cd_vinyle_createur_manquant }} as float64) * 1.0
+                cast({{ film_titre_format_support }} as int) * 1.0
+                + cast({{ absence_image }} as int) * 2.0
+                + cast({{ absence_video }} as int) * 1.0
+                + cast({{ absence_description }} as int) * 1.5
+                + cast({{ livre_cd_vinyle_artiste_manquant }} as int) * 1.0
+                + cast({{ livre_cd_vinyle_gtl_manquant }} as int) * 1.0
+                + cast({{ livre_cd_vinyle_createur_manquant }} as int) * 1.0
             )
             / 8.5
         ),
